FirebaseFunctions refactor
